+++ src/assets/config/config.ts
@@ -53,7 +53,7 @@ export const config: Config = {
     },
     highlightSearchMatches: true,
     inlineIllustrations: [206],
-    mediaCollectionMappings: { 214: 44, 206: 19 },
+    mediaCollectionMappings: { 214: 44, 206: 19, 218: 19 },
     order: [
       [216, 219, 220, 218, 210, 208, 207, 214, 203, 213,
         202, 199, 221, 206, 201, 211, 200, 205, 215, 217,